% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chromatogram_and_CSIA.R
\name{BchronDensity_plot_params}
\alias{BchronDensity_plot_params}
\title{BchronDensity_plot_params}
\usage{
BchronDensity_plot_params(ID, x)
}
\arguments{
\item{ID}{character, the site ID}

\item{x}{output object from BchronDensity}
}
\value{
A line plot for radiocarbon dates
}
\description{
BchronDensity_plot_params
}
\examples{
Dens_KM = BchronDensity(ages=dates_KM$Date, ageSds=dates_KM$Uncertainty, calCurves=rep('intcal13', nrow(dates_KM)))
BchronDensity_plot_params("KM", Dens_KM)
}

