% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chromatogram_and_CSIA.R
\name{make_retention_times_plot_no_labels}
\alias{make_retention_times_plot_no_labels}
\title{make_retention_times_plot_no_labels}
\usage{
make_retention_times_plot_no_labels(the_data)
}
\arguments{
\item{the_data}{a data.frame of two columns labeled 'retention' and 'area'}
}
\value{
A plot of retention tims by abundance showing the result of a GC-MS analysis and some compound labels
}
\description{
make_retention_times_plot_no_labels
}
\examples{
make_retention_times_plot_no_labels()
}

