% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chromatogram_and_CSIA.R
\name{make_retention_times_plot}
\alias{make_retention_times_plot}
\title{make_retention_times_plot}
\usage{
make_retention_times_plot(the_data)
}
\arguments{
\item{the_data}{a data.frame of two columns labeled 'retention' and 'area'}
}
\value{
A plot of retention tims by abundance showing the result of a GC-MS analysis and some compound labels
}
\description{
make_retention_times_plot
}
\examples{
make_retention_times_plot()
}

